Layers and spectrum builders are separated
